requirements/base.txt
Django==1.10
Python>=2.7
django-taggit==0.21.1